add registation task

+++ admin/includes/view_all_posts.php
@@ -114,7 +114,7 @@
             echo   "<td>$post_comment_count</td>";
             echo   "<td>$post_status</td>";
             echo    "<td><a href='posts.php?source=edit_post&p_id=$post_id'>Edit</a></td>";
-            echo    "<td><a href='posts.php?delete=$post_id'>Delete</a></td>";             
+            echo    "<td><a onClick=\" javascript: return confirm('What do you want delete?'); \" href='posts.php?delete=$post_id'>Delete</a></td>";             
             echo "</tr>";
         }
         ?>
